rename: 트래블로그 상세 조회 관련 폴더명 변경
