+++ src/components/TravelogueDetail/TravelInfoItem.tsx
