+++ src/components/TravelogueDetail/TravelogueHeader.tsx
