+++ src/components/TravelogueDetail/TravelogueInfo.tsx
