+++ src/components/TravelogueDetail/index.ts
@@ -1,6 +1,7 @@
 export { default as TravelInfo } from './TravelInfo';
 export { default as TravelInfoItem } from './TravelInfoItem';
 export { default as TravelogueContent } from './TravelogueContent';
+export { default as TravelogueDetail } from './TravelogueDetail';
 export { default as TravelogueHeader } from './TravelogueHeader';
 export { default as TravelogueInfo } from './TravelogueInfo';
 export { default as VisitedRegionList } from './VisitedRegionList';
