+++ src/pages/detail/index.tsx
@@ -1,15 +1,14 @@
-import { Box, Stack } from '@mui/material';
+import { Box } from '@mui/material';
 import { useRouter } from 'next/router';
 import React, { useEffect, useState } from 'react';
 
 import { usePatchTravelogueDetailById } from '@/api/hooks/travelogue';
 import Spinner from '@/components/common/Spinner';
-import { TravelogueContent, TravelogueHeader } from '@/components/PostDetail';
+import { TravelogueDetail } from '@/components/TravelogueDetail';
 import { TravelogueDetailType } from '@/types/travelogue';
 
 const Detail = () => {
   const router = useRouter();
-  const authority = 'writer'; // viewer
   const { mutate, isLoading } = usePatchTravelogueDetailById();
   const [travelogueDetail, setTravelogueDetail] = useState<TravelogueDetailType>();
 
@@ -31,12 +30,7 @@ const Detail = () => {
 
   return (
     <Box sx={{ p: '0 15px 50px 15px' }}>
-      <Stack spacing={3}>
-        <TravelogueHeader authority={authority} travelogueDetail={travelogueDetail} />
-        {travelogueDetail.subTravelogues.map((subTravelogue) => (
-          <TravelogueContent key={subTravelogue.day} subTravelogue={subTravelogue} />
-        ))}
-      </Stack>
+      <TravelogueDetail travelogueDetail={travelogueDetail} />
     </Box>
   );
 };
